moved params to separate file
